Move client discovery cache warm up to warm up stage. Incorporate code review comments.
